half-refactor most claim-related classes to claim package
