+++ src/me/ryanhamshire/GriefPrevention/FlatFileDataStore.java
@@ -20,13 +20,11 @@ package me.ryanhamshire.GriefPrevention;
 
 import java.io.*;
 import java.nio.charset.Charset;
-import java.text.DateFormat;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.regex.Matcher;
 
+import me.ryanhamshire.GriefPrevention.claim.Claim;
 import org.bukkit.*;
 import org.bukkit.configuration.InvalidConfigurationException;
 import org.bukkit.configuration.file.YamlConfiguration;
