+++ src/me/ryanhamshire/GriefPrevention/Visualization.java
@@ -20,6 +20,7 @@ package me.ryanhamshire.GriefPrevention;
 
 import java.util.ArrayList;
 
+import me.ryanhamshire.GriefPrevention.claim.Claim;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.World;
