+++ src/me/ryanhamshire/GriefPrevention/claim/ClaimPermission.java
@@ -16,7 +16,7 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
  
-package me.ryanhamshire.GriefPrevention;
+package me.ryanhamshire.GriefPrevention.claim;
 
 //basic enum stuff
 public enum ClaimPermission
